+++ CHANGES.md
@@ -4,6 +4,7 @@
  - 2025-11-06 Fixed: Images in articles are inverted in Dark Skin. Thanks for reporting to @Vocta1310 [#724](https://github.com/znuny/Znuny/issues/724).
  - 2025-11-05 Integrated package Znuny-CopyTicketNumber.
  - 2025-10-31 Fixed: ProcessPrint. When printing the process, it produces empty results.
+ - 2025-10-29 Fixed: Issue with popup redirections after submitting process category form.
  - 2025-10-29 Fixed: Parameter error in Kernel::System::Web::UploadCache does not try to dereference undef.
  - 2025-10-28 Fixed: Breadcrumb now loads immediately when switching System Configuration groups via Ajax.
  - 2025-10-21 Fixed: Issue with "Title" field on customer login screen where it wasn't set for new customer user at all.